Glue Git{hub,book} and sphinx index pages together
Just symlink all index.md -> README.md

Another route would be to set in conf.py

  root_doc = 'README'

Which works, except that then all URLs will have README.html in them;
it would work if we could configure the webserver hosting us
to treat README.html as an index page, which we can do if we self-host,
but that's probably not possible with most, and above all it's not
a very portable solution. So instead: symlinks.

Maybe there's a configuration option that can rewrite README -> index
during the build. But that'll probably mess up something else somewhere.

These were created with

find . -name "README.md" |
  while read target; do
    (
      cd $(dirname $target);
      ln -sf README.md index.md
    );
  done