/team/: tables -> lists
Lists have an easier syntax, especially since we only have one
column in these tables.

Maybe we'll have multiple columns again in the future,
but I think at that point I will be scraping Github profiles
to *generate* the content, and so prettiness will be irrelevant.
But for now it is relevant.